refactor table orderedItems added quantity

+++ Infrastructure/SqlServer/Factories/OrderedItemFactory.cs
@@ -21,7 +21,7 @@ namespace Infrastructure.SqlServer.Factories
                 {
                   Id = reader.GetInt32(reader.GetOrdinal(OrderedItemSqlServer.ColItemId)),
                   Label = reader.GetString(reader.GetOrdinal(ItemSqlServer.ColLabel)),
-                  Price = reader.GetFloat(reader.GetOrdinal(ItemSqlServer.ColPrice)),
+                  Price = (float) reader.GetDouble(reader.GetOrdinal(ItemSqlServer.ColPrice)),
                   ImageItem = reader.GetString(reader.GetOrdinal(ItemSqlServer.ColImageItem)),
                   DescriptionItem = reader.GetString(reader.GetOrdinal(ItemSqlServer.ColDescriptionItem)),
                   Category = new Category { Id = reader.GetInt32(reader.GetOrdinal(ItemSqlServer.ColCategoryId)) }
